Add a new config.utils module to hold utilities used to define config values
